CORE-254: Clean up additional TODOs and add toString behaviours

+++ Java/Crypto/src/androidTest/java/com/breadwallet/crypto/UnitAIT.java
@@ -28,7 +28,7 @@ public class UnitAIT {
         assertTrue(satoshi_btc.isCompatible(satoshi_btc));
         assertEquals(satoshi_btc.core, satoshi_btc.getBase().core);
 
-        Unit btc_btc = new Unit(btc, "BTC-BTC", "Bitcoin", "B", satoshi_btc, (byte) 8);
+        Unit btc_btc = new Unit(btc, "BTC-BTC", "Bitcoin", "B", satoshi_btc, 8);
         assertEquals(btc_btc.getCurrency().getCode(), btc.getCode());
         assertTrue(satoshi_btc.isCompatible(btc_btc));
         assertTrue(btc_btc.isCompatible(satoshi_btc));
@@ -40,7 +40,7 @@ public class UnitAIT {
         Currency eth = new Currency("Ethereum", "Ethereum", "ETH", "native");
 
         Unit satoshi_btc = new Unit(btc, "BTC-SAT", "Satoshi", "SAT");
-        Unit btc_btc = new Unit(btc, "BTC-BTC", "Bitcoin", "B", satoshi_btc, (byte) 8);
+        Unit btc_btc = new Unit(btc, "BTC-BTC", "Bitcoin", "B", satoshi_btc, 8);
         Unit wei_eth = new Unit(eth, "ETH-WEI", "WEI", "wei");
 
         assertFalse(wei_eth.isCompatible(btc_btc));
