+++ Java/Crypto/src/main/java/com/breadwallet/crypto/blockchaindb/models/bdb/CurrencyDenomination.java
@@ -19,45 +19,45 @@ public class CurrencyDenomination {
     );
 
     public static CurrencyDenomination BTC_SATOSHI = new CurrencyDenomination(
-            "satoshi", "sat", (byte) 0, lookupSymbol("sat"));
+            "satoshi", "sat", 0, lookupSymbol("sat"));
 
     public static CurrencyDenomination BTC_BITCOIN = new CurrencyDenomination(
-            "bitcoin", "btc", (byte) 8, lookupSymbol("btc"));
+            "bitcoin", "btc", 8, lookupSymbol("btc"));
 
 
     public static CurrencyDenomination BCH_BITCOIN = new CurrencyDenomination(
-            "bitcoin", "bch", (byte) 8, lookupSymbol("bch"));
+            "bitcoin", "bch", 8, lookupSymbol("bch"));
 
 
     public static CurrencyDenomination ETH_WEI = new CurrencyDenomination(
-            "wei", "wei", (byte) 0, lookupSymbol("wei"));
+            "wei", "wei", 0, lookupSymbol("wei"));
 
     public static CurrencyDenomination ETH_GWEI = new CurrencyDenomination(
-            "gwei", "gwei", (byte) 9, lookupSymbol("gwei"));
+            "gwei", "gwei", 9, lookupSymbol("gwei"));
 
     public static CurrencyDenomination ETH_ETHER = new CurrencyDenomination(
-            "ether", "eth", (byte) 18, lookupSymbol("eth"));
+            "ether", "eth", 18, lookupSymbol("eth"));
 
 
     public static CurrencyDenomination BRD_INT = new CurrencyDenomination(
-            "BRD_INTEGER", "BRDI", (byte) 0, lookupSymbol("brdi"));
+            "BRD_INTEGER", "BRDI", 0, lookupSymbol("brdi"));
 
     public static CurrencyDenomination BRD_BRD = new CurrencyDenomination(
-            "BRD", "BRD", (byte) 18, lookupSymbol("brd"));
+            "BRD", "BRD", 18, lookupSymbol("brd"));
 
 
     public static CurrencyDenomination EOS_INT = new CurrencyDenomination(
-            "EOS_INTEGER", "EOSI", (byte) 0, lookupSymbol("eosi"));
+            "EOS_INTEGER", "EOSI", 0, lookupSymbol("eosi"));
 
     public static CurrencyDenomination EOS_EOS = new CurrencyDenomination(
-            "EOS", "EOS", (byte) 18, lookupSymbol("eos"));
+            "EOS", "EOS", 18, lookupSymbol("eos"));
 
 
     public static Optional<CurrencyDenomination> asDenomination(JSONObject json) {
         try {
             String name = json.getString("name");
             String code = json.getString("short_name");
-            byte decimals = (byte) json.getInt("decimals"); // TODO: This needs to be rethought
+            int decimals = json.getInt("decimals");
             String symbol = lookupSymbol(code);
             return Optional.of(new CurrencyDenomination(name, code, decimals, symbol));
 
@@ -91,10 +91,10 @@ public class CurrencyDenomination {
 
     private final String name;
     private final String code;
-    private final byte decimals;
+    private final int decimals;
     private final String symbols;
 
-    public CurrencyDenomination(String name, String code, byte decimals, String symbols) {
+    public CurrencyDenomination(String name, String code, int decimals, String symbols) {
         this.name = name;
         this.code = code;
         this.decimals = decimals;
@@ -109,7 +109,7 @@ public class CurrencyDenomination {
         return code;
     }
 
-    public byte getDecimals() {
+    public int getDecimals() {
         return decimals;
     }
